requirements.txt
fastapi
uvicorn
grpcio
pika
requests
